Use environment variables for endpoint lists
Make new environment "the" to cache lists of environment variables/arguments/descriptions so these do not have to be scraped every time `build_query()`  is run. Unless forced they are now only scraped on the first call in a session.

Move `list_endpoint_args()` into the same file as `list_endpoints()` to associate them with the new environment